requirements.txt
pdfminer.six
nltk==3.9.1
chromadb==0.5.5
langchain==0.2.14
langchain-openai==0.1.22
langchain-chroma==0.1.2
langchain-text-splitters==0.2.2
requests==2.31.0
beautifulsoup4==4.12.3
rank-bm25==0.2.2
